requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database (SQLite for development, PostgreSQL for production)
# psycopg2-binary==2.9.7  # Uncomment for PostgreSQL production
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Payment Processing
stripe==7.8.0

# Communication & Notifications (Twilio Video + WhatsApp Business API)
twilio==8.10.3
requests==2.31.0

# File Handling & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-yasg==1.21.7

# Development & Testing
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Utilities
django-extensions==3.2.3
django-filter==23.5
djangorestframework-filters==1.0.0.dev2
python-dateutil==2.8.2
pytz==2023.3

# Video Call Integration (Twilio Video)
# Twilio SDK already included above for video calls

# WhatsApp Business API (via Twilio)
# Twilio WhatsApp API included in twilio package above

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
